Move action util functions from validation_operator module to checkpoint module
