+++ great_expectations/checkpoint/actions.py
@@ -18,11 +18,8 @@ from great_expectations.data_context.util import instantiate_class_from_config
 from great_expectations.data_context.store.metric_store import MetricStore
 from great_expectations.data_context.types.resource_identifiers import ValidationResultIdentifier
 from great_expectations.exceptions import ClassInstantiationError, DataContextError
-from great_expectations.validation_operators.util import (
-    send_microsoft_teams_notifications,
-    send_opsgenie_alert,
-    send_slack_notification,
-)
+from great_expectations.checkpoint.util import send_slack_notification, send_opsgenie_alert, \
+    send_microsoft_teams_notifications
 
 logger = logging.getLogger(__name__)
