+++ great_expectations/validation_operators/__init__.py
@@ -6,7 +6,7 @@ from .validation_operators import (
     ValidationOperator,
     WarningAndFailureExpectationSuitesValidationOperator,
 )
-from great_expectations.checkpoint.actions import (
+from great_expectations.checkpoint import (
     MicrosoftTeamsNotificationAction,
     NoOpAction,
     OpsgenieAlertAction,
